Move mocks to graphql.js
