+++ gsa/src/web/pages/tasks/graphql.js
@@ -17,6 +17,7 @@
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 import gql from 'graphql-tag';
+import {TASK_STATUS} from 'gmp/models/task';
 
 import {useMutation, useQuery} from '@apollo/react-hooks';
 
@@ -276,3 +277,97 @@ export const useCreateContainerTask = () => {
   const [createContainerTask] = useMutation(CREATE_CONTAINER_TASK);
   return toGraphQL(createContainerTask);
 };
+
+const lastReport = {
+  uuid: '1234',
+  severity: '5.0',
+  timestamp: '2020-02-27T13:20:45Z',
+};
+
+const mockTask = {
+  data: {
+    tasks: {
+      nodes: [
+        {
+          name: 'foo',
+          uuid: '1234',
+          permissions: [
+            {
+              name: 'Everything',
+            },
+          ],
+          lastReport,
+          reportCount: {
+            total: 1,
+            finished: 1,
+          },
+          status: TASK_STATUS.done,
+          target: {
+            name: 'Target',
+            uuid: 'id1',
+          },
+          trend: null,
+          comment: 'bar',
+          owner: 'admin',
+          preferences: null,
+          schedule: null,
+          alerts: [],
+          scanConfig: {
+            uuid: 'id2',
+            name: 'lorem',
+            trash: false,
+          },
+          scanner: {
+            uuid: 'id3',
+            name: 'ipsum',
+            scannerType: 'dolor',
+          },
+          hostsOrdering: null,
+          observers: {
+            users: ['john', 'jane'],
+            roles: [
+              {
+                name: 'r1',
+              },
+              {
+                name: 'r2',
+              },
+            ],
+            groups: [
+              {
+                name: 'g1',
+              },
+              {
+                name: 'g2',
+              },
+            ],
+          },
+        },
+      ],
+    },
+  },
+};
+
+export const mockGetTasks = [
+  {
+    request: {
+      query: GET_TASKS,
+      variables: {filterString: 'foo=bar rows=2'},
+    },
+    result: mockTask,
+  },
+  {
+    request: {
+      query: GET_TASKS,
+      variables: {filterString: ''},
+    },
+    result: mockTask,
+  },
+  {
+    request: {
+      query: GET_TASKS,
+      variables: {filterString: ''},
+    },
+    result: mockTask,
+  },
+];
